requirements.txt
python-dotenv==1.0.0
selenium==4.13.0
openpyxl==3.1.2